Disable trigger selection in MC
All selected triggers are marked as wasRun, but at the same time none of
them accept any event in the test sample. This is not understood.
Nonetheless since triggers were not supposed to be simulated in
background datasets, just switch off selection on them. Keep the module
in the path for the moment.