Choose trigger bin based on pt(j1) instead of pt(recoil)
The old approach is not correct since it allows in some events that fall
on the sloped region of the trigger turn-on curve. The choice of trigger
bin is also moved to a dedicated plugin.

This commit breaks compilation of the module. Other plugins and the main
program must be updated accordingly.