Change MET correction scheme to L1L2L3-RC and include JER smearing
This updates are proposed by Mikko, and they change MPF a bit (~0.5%).

Taking an advantage of improvements in the mensura framework in commits
179de32c6d7ca6000eda529182a05b6f956fb047 and
8428f90f27c89a751a8734a13fe74b7df58a6a7f, T1 MET corrections are now
applied directly on top of raw MET rather than starting from the
corrected MET and undoing old T1 corrections first. This is possible
because there are no other MET corrections and all jets are stored.

To propagate JER smearing into MET, jets are also recorrected in
simulation despite the fact that corrections stored in PEC tuples are
up-to-date. This allows the stochastic smearing in JER to be included
consistently into both jets and MET.

At the same time update JER factors in simulation to their first version
in 80X instead of using old values from 76X. JER scale factors are still
from 76X, though.