Allow overriding definitions of physics objects from PEC-tuples
At the moment all define_* functions are simply imported from
PEC-tuples, however, some of them will be replaced with local versions
in future.

Because the CMSSW subpackage is only linked insided
$CMSSW_BASE/src/Analysis, SCRAM does not add the python subdirectory
into the search path for modules. In order to fix this, one should
manually link the python subdirectory as
$CMSSW_BASE/python/Analysis/Multijet. Should probably ask for a better
solution at [1].
[1] https://hypernews.cern.ch/HyperNews/CMS/get/sw-develtools.html