Update pileup profiles and JEC
This is the first version with rederived L1 and L2 corrections up to the
end of Run2016G. Runs F and E are now considered independently since
some differences between the two have been observed.

Simulation is also split to match corresponding data-taking periods
because of small differences in L1RC corrections (which come from the
different pileup reweighting in simulation).