Switch trigger selection to matching for leading jet
Trigger selection was applied by simply checking trigger bits, while any
jet could have fired the trigger. Now require that it is the leading jet
that fires the trigger. This eliminates biases from events in which the
first two jets have similar pt, and the trigger is fired by the second
one instead of the leading jet.

The selection is implemented with trigger objects only, while stored
trigger bits are completely ignored. It has been verified that there are
no pathological situations in which an event is accepted by a trigger
according to trigger bits but contains no corresponding trigger objects,
and vice versa.

The thresholds for trigger bins are also updated (in principle, per-jet
turn-on curve is slower than per-event one). They correspond to 99%
efficiency in data and at least the same efficiency in simulation.