Change definition of beta cut and disable it by default
Minimal pt for jets considered for evaluation of beta is now computed as
a fraction of pt of the leading jet instead of the recoil. This is a
more natural definition: an event is rejected if in the vicinity of the
leading jet a jet exists whose pt is larger than the given fraction of
pt of the leading jet.

At the same time the selection on beta is disabled in the main program.
This selection is historical and not motivated well in case of CMS.