Create setup for pseudoexperiments
Pseudodata is generated by selecting some of simulated events, with the
selection probability inversely proportional to the effective
luminosities of the corresponding datasets. Only include datasets with
Ht > 1 TeV since effective luminosities of others are too small. The
target integrated luminosity is set to 5/fb.

Trigger selection is disabled for the sake of simplicity. Consider
events with ptlead > 400 GeV to respect the exclusion of low-Ht samples.
There should be no pileup reweighting.

Option "--add-jec" allows to specify an additional jet correction, which
can be used to alter jet response and see whether this modification is
then measured correctly.