Allow trigger bins overlap in pt
Change the logic of the event flow. Instead of assigning each event a
specific trigger bin based on the pt of the leading jet, probe each
event against all trigger bins and accept where it matches the allowed
pt ranges. The ranges for different trigger bins can and will overlap.
This is needed in order to allow for a migration of events between
different trigger bins as a result of L3Res jet corrections.

At the same time drop the support for multiple pt thresholds. There will
be only two, and they are not supposed to change. One threshold, 30 GeV,
is used for the pt balancing and, by extension, in the event selection
based on properties of the recoil. The second threshold, 15 GeV, is
implicit and used in the MPF method.

Program addWeights needs to be adjusted to this change as it works with
indices of trigger bins explicitly.