Eliminate polymorphism for nuisances
The design based around NuisancesBase would require dynamic up-casting
each time a deviation with a non-trivial dependence on nuisances is
evaluated. Instead, switch to a single structure to describe nuisances,
without any inheritance. It will be changed each time a nuisance
parameter is added or removed, but this would anyway call for changes in
some algorithms to compute deviations.

Conceptually, nuisances would be best described with a
map<string, double>, but multiple lookups at every evaluation of the
loss function for the fit would be wasting resources. Although this
would probably still be fast enough.