Lower threshold for jets that populate histograms in data
This allows to technically evaluate the loss function when jets migrate
from slightly below the threshold for the type 1 correction. Still, if
the jet correction allows for this, the corrected ptmiss is not
well-defined. But implementing this is considered a better compromise
than causing an exception in the loss function.