Decouple stored properties of jets and MET from PEC-tuples
Add dedicated classes to describe reconstructed jets and missing pt.
This will allow to store eventually a different set of properties than
in package PEC-tuples, which is intended for physics analysis. The
classes are simple structures without any methods and have been chosen
over integral types to save typing (one branch for all jets as opposed
to a branch with an array for every property).

Add a dedicated plugin, BasicJetMET, to store jets and missing pt using
the new classes. This plugin is a simplified version of PECJetMET used
before.

Read raw CHS MET directly from pat::MET (this fixes #13). This also
eliminates the need to recompute MET in a custom way, so MET stored in
MiniAOD can be used directly.