Drop selection on recoil
Event selection involving the recoil (especially, the cut on A) accepts
different events depending on the residual jet corrections [1-2], which
cases a systematic bias in recomputed <B>. Eliminate this selection.

Apply a loose cut on the angle between the two leading jets to reject
some of pathologically imbalanced events.

Compute balance observables with static methods in BalanceVars instead
of writing explicit definitions in multiple places.

[1] https://indico.cern.ch/event/745879/#6-inverstigation-of-multijet-n
[2] https://indico.cern.ch/event/747331/#6-l3res-independent-event-sele