Invert logic in using of margins in pt ranges for trigger bins
For data will now use the exact ranges while for simulation a margin of
around 10% will be added. Change the names of parameters in the
configuration file as there is no more a strict relation with corrected
vs uncorrected jets. In Python code only change the names mechanically;
the code will need to be changed in a non-trivial way.