Reimplement systematic variations in simulation
They are now provided in the form of splines, as needed for the fit of
multijet data with crawling bins. Old code is dropped, so the binned-sum
version of the fit will not be supported anymore.

Systematic variations in data are dropped temporarily and will need to
be reimplemented.

Relevant to issue #20.