Update L2Res corrections with varied JER
The previous version of the L2Res corrections with varied JER produced
one-sided variations. After a discussion with Anastasia, update these
corrections to a newer version and also switch to ones from the pt
balance as opposed to MPF because the latter show one-sided variations
even in the new version.