Switch plotting scripts to RDataFrame
This decrease the running time a bit. (The change is not drastic,
though; probably, because the relevant part of the input file is loaded
into memory, and the subsequent calls to TTree.Draw were relatively
fast.)

Tried enabling multithreading, but the code crashes. Haven't
investigated this. The old problem with trees in directories [1] is
supposed to have been fixed in the currently used ROOT version 6.16.

[1] https://sft.its.cern.ch/jira/browse/ROOT-9735