Update script to produce systematic variations
Align build_syst_vars.py with recent changes in the repository. In
particular, account for the drop of weight files and add support for
multiple input files.

Rework the interface to access the configuration of systematic
variations. Instead of allowing to instantiate the configuration for a
specific era at any time, construct a combination of all eras by
default, while allowing to set a specific era at the initialization.