Fix style for residuals
Undo commit [1]. While it solved the problem when the scientific
notation was used for y axis on the panel with residuals in the plots
with comparison between data and simulation, it broke some diagnostic
plots, in which using the scientific notation is advantageous.

To solve this once and for all, roll back to the previous limits in the
default style and explicitly disable scientific notation for residuals.

[1] 320dcd1d6b6605b75c9789e6f102cff65b726887